Model repo refactor and fixes (#396)
* main should not use code from react

* use types from main world

* make model repo call from client directly
