+++ public/interfaces/eventHubInterface.ts
@@ -2,7 +2,6 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License
  **********************************************************/
-import { Message as MessageInterface } from '../../src/app/api/models/messages';
 export interface StartEventHubMonitoringParameters {
     deviceId: string;
     consumerGroup: string;
@@ -14,7 +13,12 @@ export interface StartEventHubMonitoringParameters {
     hubConnectionString?: string;
 }
 
-export type Message = MessageInterface;
+export interface Message {
+    body: any; // tslint:disable-line:no-any
+    enqueuedTime: string;
+    properties?: any; // tslint:disable-line:no-any
+    systemProperties?: {[key: string]: string};
+}
 
 export interface EventHubInterface {
     startEventHubMonitoring(params: StartEventHubMonitoringParameters): Promise<Message[]>;
